requirements.txt
pandas
aiohttp
psycopg2-binary
pytest